move packet_util.py to utils folder and fix imports
