+++ serial_test/iliad_data_controller.py
@@ -1,7 +1,7 @@
-import serial_data_controller
+from . import serial_data_controller
 import serial
 import struct
-import packet_util
+from utils import packet_util
 import crc
 
 class IliadDataController(serial_data_controller.SerialDataController):
